fix: SavePreferences being called multiple times and causing issues
